Update webhook-request-parser.md - Round #1
I noticed quite a few places where we were coding in the video, but the code wasn't being addressed in the audio, so just make sure whatever was left out was intentional. I added a little bit back in to fill time, otherwise we just have a bunch of silent sections where we're coding and not talking about it.