Update list-orders.md - Round #1
I made some MAJOR changes to this one. I cut out the entire section dealing with that workaround that ultimately didn't work, and I came across several sections where the code didn't match what was on the screen, or where we did something on the screen, but didn't address it in the script. 

This entire video needs a re-record after the content of the script is accurate and approved. Hopefully that will help cut down on time as well, because 31 minutes is excessive.